uFunc can now take the optional parameter iterable_input. If True, it will map the function f to all elements in the input, assuming they are iterables, instead of recursively looking for non-iterable inputs to apply f. Note: in the future another parameter may be added to control the depth to which the algorithm should get until it stops.
Added physical constants to science modules along with quantum gates and tools for
working with quantum computation. d functions found in physics is now called distance,
and was updated to work with many spatial velocities along with a custom metric tensor
(minkowski metric by default).
Different coordinate velocities are passed as a tuple, and the metric tensor should be
a Matrix or np.array.
time_dilation function was also updated for working with many spatial dimensions.
Note: when using distance() or time_dilation(), time coordinate, t, must be the first
element of the tuple. Also, it should be measured in light seconds and not seconds.
The latter conversion is done by multipying by c.

Vectors and Matrices can now be applied to len() returning dimension and m_dimension
respectively. Vector 1-dimensional input was deprecated. For correct functioning, it must be
2-dimensional. e.g:

	>>> Vector([1,2,4])   #Wrong way
	>>> Vector([[1,2,4]]) #Right way

The same must be applied when inputting a np.array