Added type checking to some functions and classes.
Cython will be used for some functions in MathiPy, like trigonometric
or exponential functions in order to improve code efficiency and speed.

Added cfuncs directory, containing source files for compiling cython
code:
	Inside src there can be found pyx files and c files that are
	compiled and imported in other MathiPy modules

	binary files are located inside bin directory. Shared object
	files (.so) are built with cython, from the setup.py file inside
	cfuncs or with g++. The command used for building with cython is:
		$ python3 setup.py build_ext

	for building with g++:
		$ g++ -fPIC -shared -o <libName.so> name.cpp

	In the build directory, .o extension files are located, created
	in the cython compilation process or with gcc.

	setup.py handles build process for cython extensions.

	headers directory contains the headers for the c files found in
	src/c

Many functions defined in _math are implemented in c or were cythonized.
The idea is to implement fast algorithms for trignonometric and
exponential functions like:
	sin, cos, tan
	arcsin, arccos, arctan
	exp, ln, root

root function uses the Newton-Raphson algorithm. The rest use Taylor
polynomials for calculations.

cython_utils.c is used for defining tools used from the .pyx files,
like poly_eval and poly_eval_w_parity used for evaluating the Taylor
polynomials, or magnitude_order, used to get the order of magnitude of
a number.

mpy_c_utils.cpp has the aim of containing functions that are not compiled
by cython, but still is compiled as a shared object to be imported.
However, performance has for now been worse than expected, so it will be
modified in the future, or removed if it still does not work as expected

algebra.py was removed from the project. Only set_contains and set_is_contained
will be kept.

logic.py was added to science submodule. xor, implication and biconditional
logic operators are defined within this file. Since algebra.py was removed,
set_contains and set_is_contained functions were moved to this file.

New functions were added to numeric_operations.py:
	ceil, floor
	trunc, mantissa
	is_integer
	get_error
	round_significant_figures
And to _math.py:
	index_of_fib
	fibonacci_seq
	sci_notation

Measure object was implemented in physics submodule. Measures in real life
are used for representing range of values as a result of a physical measure.
The Measure object it is used then for operating with this range of values
with their own arithmetic rules.

Some classes now carry the dataclass decorator from the dataclasses module.

testing submodule was removed. get_error function was moved to numeric_operations.py

config.py was added to mathipy. Both functions defined inside of it by
the moment are used to get and import the functions from mpy_c_utils.cpp.
But, since the latter is not delivering the expected results, config.py
might be also modified.

requirements.txt was updated, now containing Cython as requirement.