Mathipy Update 24/09/21
Minor updates.
Added requirements_dev.

linalg submodule will be migrated to a more functional
module, and instead of offering the Vector and Matrix classes,
it will contain functions that work with NumPy ndarrays.
The idea is to create a more easy to integrate and use module,
since it depends, and therefore is compatible with, on numpy.